py-fast-api/requirements.txt
fastapi
psycopg2
uvicorn
python-dotenv
redis
google-generativeai